Move state-key to separate module
