+++ packages/headless/src/app/state-key.ts
@@ -1,0 +1,1 @@
+export const stateKey = Symbol('state');
