+++ packages/headless/src/controllers/commerce/context/cart/headless-cart.ts
@@ -1,6 +1,6 @@
 import {CurrencyCodeISO4217, Ec} from '@coveo/relay-event-types';
 import {CommerceEngine} from '../../../../app/commerce-engine/commerce-engine';
-import {stateKey} from '../../../../app/engine';
+import {stateKey} from '../../../../app/state-key';
 import {
   purchase,
   setItems,
