+++ packages/headless/src/controllers/commerce/context/headless-context.ts
@@ -1,6 +1,6 @@
 import {CurrencyCodeISO4217} from '@coveo/relay-event-types';
 import {CommerceEngine} from '../../../app/commerce-engine/commerce-engine';
-import {stateKey} from '../../../app/engine';
+import {stateKey} from '../../../app/state-key';
 import {
   setContext,
   setUser,
