+++ packages/headless/src/controllers/commerce/core/facets/generator/headless-commerce-facet-generator.ts
@@ -3,7 +3,7 @@ import {
   CommerceEngine,
   CommerceEngineState,
 } from '../../../../../app/commerce-engine/commerce-engine';
-import {stateKey} from '../../../../../app/engine';
+import {stateKey} from '../../../../../app/state-key';
 import {commerceFacetSetReducer as commerceFacetSet} from '../../../../../features/commerce/facets/facet-set/facet-set-slice';
 import {CommerceFacetSetState} from '../../../../../features/commerce/facets/facet-set/facet-set-state';
 import {FacetType} from '../../../../../features/commerce/facets/facet-set/interfaces/common';
