+++ packages/headless/src/controllers/commerce/core/facets/numeric/headless-commerce-numeric-facet.ts
@@ -1,5 +1,5 @@
 import {CommerceEngine} from '../../../../../app/commerce-engine/commerce-engine';
-import {stateKey} from '../../../../../app/engine';
+import {stateKey} from '../../../../../app/state-key';
 import {
   toggleExcludeNumericFacetValue,
   toggleSelectNumericFacetValue,
