+++ packages/headless/src/controllers/commerce/core/facets/regular/headless-commerce-regular-facet-search.test.ts
@@ -1,4 +1,4 @@
-import {stateKey} from '../../../../../app/engine';
+import {stateKey} from '../../../../../app/state-key';
 import {
   executeCommerceFacetSearch,
   executeCommerceFieldSuggest,
