+++ packages/headless/src/controllers/commerce/core/facets/regular/headless-commerce-regular-facet-search.ts
@@ -1,6 +1,6 @@
 import {SpecificFacetSearchResult} from '../../../../../api/search/facet-search/specific-facet-search/specific-facet-search-response';
 import {CommerceEngine} from '../../../../../app/commerce-engine/commerce-engine';
-import {stateKey} from '../../../../../app/engine';
+import {stateKey} from '../../../../../app/state-key';
 import {
   executeCommerceFacetSearch,
   executeCommerceFieldSuggest,
