+++ packages/headless/src/controllers/commerce/core/facets/regular/headless-commerce-regular-facet.ts
@@ -3,7 +3,7 @@ import {
   CommerceEngine,
   CommerceEngineState,
 } from '../../../../../app/commerce-engine/commerce-engine';
-import {stateKey} from '../../../../../app/engine';
+import {stateKey} from '../../../../../app/state-key';
 import {
   toggleExcludeFacetValue,
   toggleSelectFacetValue,
