+++ packages/headless/src/controllers/commerce/core/pagination/headless-core-commerce-pagination.test.ts
@@ -1,4 +1,4 @@
-import {stateKey} from '../../../../app/engine';
+import {stateKey} from '../../../../app/state-key';
 import {
   selectPage,
   nextPage,
