+++ packages/headless/src/controllers/commerce/core/pagination/headless-core-commerce-pagination.ts
@@ -4,7 +4,7 @@ import {
   CommerceEngine,
   CommerceEngineState,
 } from '../../../../app/commerce-engine/commerce-engine';
-import {stateKey} from '../../../../app/engine';
+import {stateKey} from '../../../../app/state-key';
 import {
   nextPage,
   previousPage,
