+++ packages/headless/src/controllers/commerce/core/result-list/headless-core-interactive-result.ts
@@ -3,7 +3,7 @@ import {
   CommerceEngine,
   CommerceEngineState,
 } from '../../../../app/commerce-engine/commerce-engine';
-import {stateKey} from '../../../../app/engine';
+import {stateKey} from '../../../../app/state-key';
 import {productClick} from '../../../../features/commerce/context/product/product-actions';
 import {
   buildInteractiveResultCore,
