+++ packages/headless/src/controllers/commerce/core/sort/headless-core-commerce-sort.ts
@@ -1,6 +1,6 @@
 import {Schema} from '@coveo/bueno';
 import {CommerceEngine} from '../../../../app/commerce-engine/commerce-engine';
-import {stateKey} from '../../../../app/engine';
+import {stateKey} from '../../../../app/state-key';
 import {
   buildFieldsSortCriterion,
   buildRelevanceSortCriterion,
