+++ packages/headless/src/controllers/commerce/core/url-manager/headless-core-url-manager.test.ts
@@ -1,4 +1,4 @@
-import {stateKey} from '../../../../app/engine';
+import {stateKey} from '../../../../app/state-key';
 import {
   CommerceSearchParameters,
   restoreSearchParameters,
