+++ packages/headless/src/controllers/commerce/core/url-manager/headless-core-url-manager.ts
@@ -1,5 +1,5 @@
 import {CommerceEngine} from '../../../../app/commerce-engine/commerce-engine';
-import {stateKey} from '../../../../app/engine';
+import {stateKey} from '../../../../app/state-key';
 import {Parameters} from '../../../../features/commerce/search-parameters/search-parameter-actions';
 import {Serializer} from '../../../../features/commerce/search-parameters/search-parameter-serializer';
 import {deepEqualAnyOrder} from '../../../../utils/compare-utils';
