+++ packages/headless/src/controllers/commerce/recommendations/headless-recommendations.ts
@@ -5,7 +5,7 @@ import {
   CommerceEngine,
   CommerceEngineState,
 } from '../../../app/commerce-engine/commerce-engine';
-import {stateKey} from '../../../app/engine';
+import {stateKey} from '../../../app/state-key';
 import {recommendationsOptionsSchema} from '../../../features/commerce/recommendations/recommendations';
 import {
   fetchRecommendations,
