+++ packages/headless/src/controllers/commerce/search/breadcrumb-manager/headless-search-breadcrumb-manager.test.ts
@@ -1,4 +1,4 @@
-import {stateKey} from '../../../../app/engine';
+import {stateKey} from '../../../../app/state-key';
 import {
   AnyFacetResponse,
   RegularFacetValue,
