+++ packages/headless/src/controllers/commerce/standalone-search-box/headless-standalone-search-box.test.ts
@@ -1,5 +1,5 @@
 import {configuration} from '../../../app/common-reducers';
-import {stateKey} from '../../../app/engine';
+import {stateKey} from '../../../app/state-key';
 import {updateQuery} from '../../../features/commerce/query/query-actions';
 import {queryReducer as commerceQuery} from '../../../features/commerce/query/query-slice';
 import {
