+++ packages/headless/src/controllers/commerce/standalone-search-box/headless-standalone-search-box.ts
@@ -1,6 +1,6 @@
 import {CommerceEngine} from '../../../app/commerce-engine/commerce-engine';
 import {configuration} from '../../../app/common-reducers';
-import {stateKey} from '../../../app/engine';
+import {stateKey} from '../../../app/state-key';
 import {updateQuery} from '../../../features/commerce/query/query-actions';
 import {queryReducer as commerceQuery} from '../../../features/commerce/query/query-slice';
 import {
